requirements.txt
faiss-cpu==1.11.0
numpy==2.3.0
pandas==2.3.0
chromadb==1.0.13